Fix build from sdist
PR #24 deduplicated the requirements, but this makes the sdist on PyPi
fail to build, because requirements.txt is not included.

This fix includes requirements.txt using MANIFEST.in and changes the
publication workflow to use build instead of calling setup.py directly.
build always builds the wheel from the sdist, so this error is
prevented in the future.